pyproject.toml
[tool.poetry]
name = "callensights"
version = "0.1.0"
description = "Callensights API Backend"
authors = ["Chandrashekhar R <shekhar.rachakonda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
boto3 = "^1.34.58"
pyjwt = "^2.8.0"
pymongo = "^4.6.2"
mysql-connector-python = "^8.3.0"
sqlalchemy = "^2.0.28"
bcrypt = "^4.1.2"
rsa = "^4.9"
passlib = "^1.7.4"
pydantic-settings = "^2.2.1"
watchtower = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
